adding image search and moving to original module
