+++ frb/surveys/wise.py
@@ -1,37 +1,40 @@
 """WISE Survey"""
 
-from IPython import embed
-
-import numpy as np
 from astropy import units, io, utils
 
-from frb.surveys import dlsurvey
+from frb.surveys import surveycoord
 from frb.surveys import catalog_utils
+from astropy.table import Table
 
-# Dependencies
 try:
-    from pyvo.dal import sia
+    from pyvo.dal import TAPService, sia
 except ImportError:
-    print("Warning:  You need to install pyvo to retrieve DES images")
+    print("Warning: You need to install pyvo to use the survey tools...")
     _svc = None
 else:
-    _DEF_ACCESS_URL = "https://datalab.noao.edu/sia/des_dr1"
+    _DEF_ACCESS_URL = "https://irsa.ipac.caltech.edu/ibe/data/wise/allwise/p3am_cdd"
     _svc = sia.SIAService(_DEF_ACCESS_URL)
 
 # Define the data model for WISE data
 photom = {}
 
-# DES-WISE
+# Define the default set of query fields
+# See: https://wise2.ipac.caltech.edu/docs/release/allwise/expsup/sec2_1a.html
+# for additional Fields
 photom['WISE'] = {}
 WISE_bands = ['W1', 'W2', 'W3', 'W4']
+_DEFAULT_query_fields = ['source_id','ra','dec','tmass_key']
 for band in WISE_bands:
-    photom['WISE']['WISE_{:s}'.format(band)] = '{:s}mpro'.format(band.lower())
-    photom['WISE']['WISE_{:s}_err'.format(band)] = '{:s}sigmpro'.format(band.lower())
+    magcol = '{:s}mag'.format(band.lower()) # The "standard" aperture measurement
+    errcol = '{:s}sigm'.format(band.lower())
+    photom['WISE']['WISE_{:s}'.format(band)] = magcol
+    photom['WISE']['WISE_{:s}_err'.format(band)] = errcol
+    _DEFAULT_query_fields.append(magcol)
+    _DEFAULT_query_fields.append(errcol)
 photom['WISE']['ra'] = 'ra'
 photom['WISE']['dec'] = 'dec'
 
-
-class WISE_Survey(dlsurvey.DL_Survey):
+class WISE_Survey(surveycoord.SurveyCoord):
     """
     Class to handle queries on the WISE survey
 
@@ -44,14 +47,14 @@ class WISE_Survey(dlsurvey.DL_Survey):
     """
 
     def __init__(self, coord, radius, **kwargs):
-        dlsurvey.DL_Survey.__init__(self, coord, radius, **kwargs)
+        surveycoord.SurveyCoord.__init__(self, coord, radius, **kwargs)
         self.survey = 'WISE'
-        self.bands = ['W1', 'W2', 'W3', 'W4']
-        self.svc = sia.SIAService("https://datalab.noao.edu/sia/allwise")
-        self.qc_profile = "default"
-        self.database = "allwise.source"
+        self.bands = WISE_bands
+        self.service = TAPService('https://irsa.ipac.caltech.edu/TAP')
+        self.query = None
+        self.database = "allwise_p3as_psd"
 
-    def get_catalog(self, query=None, query_fields=None, print_query=False):
+    def get_catalog(self, query=None, query_fields=_DEFAULT_query_fields, print_query=False):
         """
         Grab a catalog of sources around the input coordinate to the search radius
 
@@ -64,20 +67,52 @@ class WISE_Survey(dlsurvey.DL_Survey):
             astropy.table.Table:  Catalog of sources returned.  Includes WISE
             photometry for matched sources.
         """
-        # Main DES query
-        main_cat = super(WISE_Survey, self).get_catalog(query_fields=query_fields, print_query=print_query)
+        # Main WISE query
+        if query is None:
+            self._gen_cat_query(query_fields)
+        if print_query:
+            print(self.query)
+        main_cat = self.service.run_async(self.query).to_table()
+        main_cat.meta['radius'] = self.radius
+        main_cat.meta['survey'] = self.survey
         if len(main_cat) == 0:
+            main_cat = catalog_utils.clean_cat(main_cat, photom['WISE'], fill_mask=-999.)
             return main_cat
+        
         main_cat = catalog_utils.clean_cat(main_cat, photom['WISE'], fill_mask=-999.)
 
         # Finish
         self.catalog = main_cat
         self.validate_catalog()
-        return self.catalog
-
-    def _gen_cat_query(self,query_fields=None):
+        return self.catalog.copy()
+    
+    def get_cutout(self, imsize, filter, timeout=120):
         """
-        Generate SQL query for catalog search
+        Download an image from IRSA
+        Args:
+            imsize(Quantity): Size of the cutout in angular units.
+            filter(str): One of "W1", "W2", "W3" or "W4"
+            timeout(float): Number of seconds to wait to hear a response from
+                the IRSA SIA server.
+        Returns:
+            imghdu(fits.HDU): Fits HDU with image
+        """
+        assert filter.upper() in self.bands, "Invalid filter name "+filter
+        # First get a table with the image coadd_id
+        meta_url = "https://irsa.ipac.caltech.edu/ibe/search/wise/allsky/4band_p3am_cdd?POS={:f},{:f}".format(self.coord.ra.value, self.coord.dec.value)
+        img_metatab = Table.read(utils.data.download_file(meta_url, cache=True, show_progress=False, timeout=timeout), format="ascii")
+        coadd_id = img_metatab['coadd_id'][img_metatab['band']==int(filter[1])][0]
+
+        # Now generate the image url
+        img_url = "https://irsa.ipac.caltech.edu/ibe/data/wise/allsky/4band_p3am_cdd/{:s}/{:s}/{:s}/{:s}-{:s}-int-3.fits".format(coadd_id[:2], coadd_id[:4], coadd_id, coadd_id, filter.lower())
+        img_url +="?center={:f},{:f}&size={:f}pix".format(self.coord.ra.value, self.coord.dec.value, imsize.to("arcsec").value/2.75)
+        self.cutout = io.fits.open(utils.data.download_file(img_url,cache=True,show_progress=False,timeout=timeout))[0]
+        self.cutout_size = imsize
+        return self.cutout.copy()
+        
+    def _gen_cat_query(self,query_fields=_DEFAULT_query_fields):
+        """
+        Generate ADQL query for catalog search
 
         self.query is modified in place
 
@@ -85,33 +120,13 @@ class WISE_Survey(dlsurvey.DL_Survey):
             query_fields (list):  Override the default list for the SQL query
 
         """
-        if query_fields is None:
-            object_id_fields = ['source_id','ra','dec','tmass_key']
-            mag_fields = ['w1mpro', 'w2mpro', 'w3mpro', 'w4mpro',
-                          'w1sigmpro', 'w2sigmpro', 'w3sigmpro', 'w4sigmpro', 'ph_qual',
-                          'moon_lev']
-            query_fields = object_id_fields+mag_fields
-
-        self.query = dlsurvey._default_query_str(query_fields, self.database, self.coord, self.radius)
-
-    def _select_best_img(self,imgTable,verbose,timeout=120):
-        """
-        Select the best band for a cutout
-
-        Args:
-            imgTable: Table of images
-            verbose (bool):  Print status
-            timeout (int or float):  How long to wait before timing out, in seconds
-
-        Returns:
-            HDU: header data unit for the downloaded image
-
-        """
-        row = imgTable[np.argmax(imgTable['exptime'].data.data.astype('float'))] # pick image with longest exposure time
-        url = row['access_url'].decode()
-        if verbose:
-            print ('downloading deepest stacked image...')
-
-        imagedat = io.fits.open(utils.data.download_file(url,cache=True,show_progress=False,timeout=timeout))
-        return imagedat
-
+        query_field_str = ""
+        for field in query_fields:
+            query_field_str += " {:s},".format(field)
+        # Remove last comma
+        query_field_str = query_field_str[:-1]
+        self.query = """SELECT{:s}
+        FROM {:s}
+        WHERE CONTAINS(POINT('ICRS',ra, dec), CIRCLE('ICRS',{:f},{:f},{:f}))=1""".format(query_field_str,self.database,self.coord.ra.value,
+                                self.coord.dec.value,self.radius.to(units.deg).value)
+        return self.query
